Update readme.md
I do not know why the two links which point to the title of two parts in the very begining can not work.  And also I do not know whether the photos in this document can display or not after my fork.I have tried my best to make it look like a tutorial.  But there are something maybe will make students confused since they are just not familiar with Andorid.  I test the code you write in this document, it can work well.  but I think you'd better mention that the class DownloadUpdate should be written in the class Mainactivity not only in the file MainActivity.java and they should creat a object of class DownloadUpdate in the function of click and implement the methord execute...What's more, I think your student will very appreciate it if you can explain the theory of AsyncTask sush as when doInBackground finish its work, the return value of it will be given to onPostExecute and execute onPostExecute.